+++ yandex/resource_yandex_mdb_clickhouse_cluster.go
@@ -676,6 +676,7 @@ func resourceYandexMDBClickHouseCluster() *schema.Resource {
 			},
 			"cloud_storage": {
 				Type:     schema.TypeList,
+				Computed: true,
 				MaxItems: 1,
 				Optional: true,
 				Elem: &schema.Resource{